There is Two options of drop panel InvectorySelector + default InvectorySelector only to display info. For users with permission stonedrop.drop + Advanced InvectorySelectorAdmin with existing possibility of Secondwindow, switch drop and stack drop
