--- src/main/java/me/apisek12/StoneDrop/InventorySelectors/InventorySelectorUser.java
@@ -1,5 +1,0 @@
-package me.apisek12.StoneDrop.InventorySelectors;
-
-public class InventorySelectorUser extends  InventorySelector{
-
-}
